Update JamfConnect_2_uninstall.sh
Added if-then tests to check for existence of files we are removing. Also want the script to exit 0 and redirect stderr to stdout.

+++ JamfConnect_2_uninstall.sh
@@ -59,26 +59,32 @@ alias sudo=/usr/bin/sudo
 
 # Uninstalling the Login Window
 
-# Reset the authentication database by executing the following command with the authchanger:
-/usr/local/bin/authchanger -reset
+if [[ -e /usr/local/bin/authchanger ]]
+then
+	# Reset the authentication database by executing the following command with the authchanger:
+	/usr/local/bin/authchanger -reset
 
-# Important: If you do not reset the authentication database before deleting Jamf Connect Login files, users will be unable to log in.
+	# Important: If you do not reset the authentication database before deleting Jamf Connect Login files, users will be unable to log in.
 
-#Remove the following files installed with Jamf Connect Login by executing the following commands:
-rm /usr/local/bin/authchanger
-rm /usr/local/lib/pam/pam_saml.so.2
-rm -r /Library/Security/SecurityAgentPlugins/JamfConnectLogin.bundle
+	#Remove the following files installed with Jamf Connect Login by executing the following commands:
+	rm /usr/local/bin/authchanger  2>&1
+	rm /usr/local/lib/pam/pam_saml.so.2 2>&1
+	rm -r /Library/Security/SecurityAgentPlugins/JamfConnectLogin.bundle 2>&1
+fi
 
 # Uninstalling the Menu Bar App
 
 # Remove the Jamf Connect app from /Applicatons 
-rm -fR "/Applications/Jamf Connect.app"
+if [[ -e "/Applications/Jamf Connect.app" ]]
+then
+	rm -fR "/Applications/Jamf Connect.app"
+fi
 
 # Remove the Jamf Connect launch agent from /Library/LaunchAgents
 if [[ -f "/Library/LaunchAgents/com.jamf.connect.plist" ]]
 then
-	/bin/launchctl unload -wF "/Library/LaunchAgents/com.jamf.connect.plist"
-	rm -f "/Library/LaunchAgents/com.jamf.connect.plist"
+	/bin/launchctl unload -wF "/Library/LaunchAgents/com.jamf.connect.plist" 2>&1
+	rm -f "/Library/LaunchAgents/com.jamf.connect.plist" 2>&1
 fi
 
-exit
+exit 0
